AbstractErrorPayloadUnionTypeResolver
