+++ layers/CMSSchema/packages/custompost-mutations/src/TypeResolvers/UnionType/CustomPostUpdateMutationErrorPayloadUnionTypeResolver.php
@@ -5,15 +5,12 @@ declare(strict_types=1);
 namespace PoPCMSSchema\CustomPostMutations\TypeResolvers\UnionType;
 
 use PoPCMSSchema\CustomPostMutations\RelationalTypeDataLoaders\UnionType\CustomPostUpdateMutationErrorPayloadUnionTypeDataLoader;
-use PoPSchema\SchemaCommons\TypeResolvers\InterfaceType\IsErrorPayloadInterfaceTypeResolver;
+use PoPSchema\SchemaCommons\TypeResolvers\UnionType\AbstractErrorPayloadUnionTypeResolver;
 use PoP\ComponentModel\RelationalTypeDataLoaders\RelationalTypeDataLoaderInterface;
-use PoP\ComponentModel\TypeResolvers\InterfaceType\InterfaceTypeResolverInterface;
-use PoP\ComponentModel\TypeResolvers\UnionType\AbstractUnionTypeResolver;
 
-class CustomPostUpdateMutationErrorPayloadUnionTypeResolver extends AbstractUnionTypeResolver
+class CustomPostUpdateMutationErrorPayloadUnionTypeResolver extends AbstractErrorPayloadUnionTypeResolver
 {
     private ?CustomPostUpdateMutationErrorPayloadUnionTypeDataLoader $customPostUpdateMutationErrorPayloadUnionTypeDataLoader = null;
-    private ?IsErrorPayloadInterfaceTypeResolver $isErrorPayloadInterfaceTypeResolver = null;
 
     final public function setCustomPostUpdateMutationErrorPayloadUnionTypeDataLoader(CustomPostUpdateMutationErrorPayloadUnionTypeDataLoader $customPostUpdateMutationErrorPayloadUnionTypeDataLoader): void
     {
@@ -23,15 +20,6 @@ class CustomPostUpdateMutationErrorPayloadUnionTypeResolver extends AbstractUnio
     {
         /** @var CustomPostUpdateMutationErrorPayloadUnionTypeDataLoader */
         return $this->customPostUpdateMutationErrorPayloadUnionTypeDataLoader ??= $this->instanceManager->getInstance(CustomPostUpdateMutationErrorPayloadUnionTypeDataLoader::class);
-    }
-    final public function setIsErrorPayloadInterfaceTypeResolver(IsErrorPayloadInterfaceTypeResolver $isErrorPayloadInterfaceTypeResolver): void
-    {
-        $this->isErrorPayloadInterfaceTypeResolver = $isErrorPayloadInterfaceTypeResolver;
-    }
-    final protected function getIsErrorPayloadInterfaceTypeResolver(): IsErrorPayloadInterfaceTypeResolver
-    {
-        /** @var IsErrorPayloadInterfaceTypeResolver */
-        return $this->isErrorPayloadInterfaceTypeResolver ??= $this->instanceManager->getInstance(IsErrorPayloadInterfaceTypeResolver::class);
     }
 
     public function getTypeName(): string
@@ -47,15 +35,5 @@ class CustomPostUpdateMutationErrorPayloadUnionTypeResolver extends AbstractUnio
     public function getRelationalTypeDataLoader(): RelationalTypeDataLoaderInterface
     {
         return $this->getCustomPostUpdateMutationErrorPayloadUnionTypeDataLoader();
-    }
-
-    /**
-     * @return InterfaceTypeResolverInterface[]
-     */
-    public function getUnionTypeInterfaceTypeResolvers(): array
-    {
-        return [
-            $this->getIsErrorPayloadInterfaceTypeResolver(),
-        ];
     }
 }
